bcs-ui/pyproject.toml
[tool.poetry]
authors = ["blueking"]
description = "blueking BCS SaaS project"
name = "bk-bcs-saas"
version = "0.0.1"

[tool.poetry.dependencies]
arrow = "==0.10.0"
bleach = "==2.0.0"
boto = "==2.49.0"
# Manually lock boto3 packge version
boto3 = "==1.7.84"
# botocore = "==1.10.84"
celery = "==3.1.18"
cryptography = "==3.4.7"
dataclasses = "^0.7"
django = "==3.2.2"
django-celery = "==3.3.1"
django-cors-headers = "==3.7.0"
django-extensions = "==3.1.3"
django-picklefield = "==3.0.1"
django-redis = "==4.12.1"
djangorestframework = "==3.12.4"
"dogpile.cache" = "==0.6.4"
dpath = "==1.4.2"
furl = "==2.0.0"
gevent = "==1.5.0"
gunicorn = "==19.7.1"
jinja2 = "==2.10.1"
jsonfield = "==3.1.0"
jsonschema = "==2.6.0"
kubernetes = "==17.17.0"
mako = "==1.0.7"
natsort = "==5.4.1"
openpyxl = "==3.0.0"
pillow = "==5.1.0"
pycrypto = "==2.6.1"
pyjwt = "==1.5.3"
pymysql = "==1.0.2"
python = ">=3.6.1,<3.7"
python-dateutil = "==2.6.1"
python-logstash = "==0.4.6"
pytz = "==2021.1"
pyyaml = "==5.3.1"
raven = "==6.10.0"
redis = "==3.5.3"
requests = "==2.21.0"
"ruamel.yaml" = "==0.15.79"
"ruamel.yaml.cmd" = "==0.5.3"
semantic-version = "2.8.4"
six = "==1.16.0"
tornado = "==5.1.1"
werkzeug = "==0.15.3"
whitenoise = "==5.2.0"
bk-iam = "1.1.9"
attrs = "^20.3.0"
aiohttp = "^3.7.4"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
channels = "^3.0.3"
wrapt = "1.12.1"

[tool.poetry.dev-dependencies]
coverage = "4.5.3"
flake8 = "3.7.8"
ipdb = "0.10.3"
ipython = "6.1.0"
mock = "3.0.5"
pytest = "6.2.4"
pytest-django = "3.9.0"
pytest-asyncio = "^0.15.1"
requests-mock = "1.9.3"

[tool.black]
exclude = '''
/(
  .*/migrations
)/
'''
line-length = 119
skip-string-normalization = 'true'

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
multi_line_output = 3
use_parentheses = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
